feat(cookies-dialog): move to components
